fix bug

+++ subgraph/greedy_expand.py
@@ -28,7 +28,7 @@ def numba_greedy_grow_tree(num_neighbors: int, weights: np.ndarray, edge_index: 
 
 def greedy_grow_tree(graph: Data, num_neighbors: int, weights: torch.Tensor, target_dtype=torch.float) -> torch.Tensor:
     if num_neighbors >= graph.num_nodes:
-        return torch.ones_like(weights, device=weights.device, dtype=torch.float)
+        return torch.ones_like(weights, device=weights.device, dtype=target_dtype)
 
     np_mask = numba_greedy_grow_tree(num_neighbors,
                                      weights.cpu().detach().numpy(),
